RenderBase, defined basic interface
Commit remaining files of the render base API.
We can start experimenting with some render pipelines using
the render base API, and start writing tests against it.

+++ Ludens/Core/RenderBase/Include/RTexture.h
@@ -21,8 +21,8 @@ namespace LD {
 	{
 		RTextureType Type;
 		RTextureFormat Format;
-		u16 Width;
-		u16 Height;
+		u32 Width;
+		u32 Height;
 		const u8* Data;
 	};
